refactored sanitize filename and it's tests
