+++ app/helpers/repository_helper.rb
@@ -1,5 +1,5 @@
 module RepositoryHelper
-  include SubmissionsHelper
+  include FileHelper
 
   # Add new files or overwrite existing files in this open +repo+. +f+ should be a
   # ActionDispatch::Http::UploadedFile object, +user+ is the user that is responsible for
