+++ spec/helpers/submission_helper_spec.rb
@@ -1,42 +1,4 @@
 describe SubmissionsHelper do
-
-  # Put some confidence in our submission filename sanitization
-  context 'A new file when submitted' do
-    context "containing characters outside what's allowed in a filename" do
-      before :each do
-        @filenames_to_be_sanitized = [ { expected: 'llll_', orig: 'llllé' },
-                                       { expected: '________', orig: 'öä*?`ßÜÄ' },
-                                       { expected: '', orig: nil },
-                                       { expected: 'garbage-__.txt', orig: 'garbage-éæ.txt' },
-                                       { expected: 'space_space.txt', orig: 'space space.txt' },
-                                       { expected: '______.txt', orig: '      .txt' },
-                                       { expected: 'garbage-__.txt', orig: 'garbage-éæ.txt' } ]
-      end
-
-      it 'have sanitized them properly' do
-        @filenames_to_be_sanitized.each do |item|
-          expect(sanitize_file_name(item[:orig])).to eq item[:expected]
-        end
-      end
-    end
-
-    context 'containing only valid characters in a filename' do
-      before :each do
-        @filenames_not_to_be_sanitized = %w(valid_file.sh
-                                            valid_001.file.ext
-                                            valid-master.png
-                                            some__file___.org-png
-                                            001.txt)
-      end
-
-      it 'NOT have sanitized away any of their characters' do
-        @filenames_not_to_be_sanitized.each do |orig|
-          expect(sanitize_file_name(orig)).to eq orig
-        end
-      end
-    end
-  end
-
   describe '#get_file_info' do
     let(:assignment) { create(:assignment) }
     let(:grouping) { create(:grouping_with_inviter_and_submission, assignment: assignment) }
